Invert api/client.ts's upward import of hooks/cachedResource (#2893)
api/client.ts needs dropCachesAfterWrite on every mutating request, which
made an api module import a hook. The React-free cache mechanics
(createResourceCache, createKeyedResourceCache, isFresh,
dropCachesAfterWrite, the TTL classes) move to the new
utils/resourceCache.ts; hooks/cachedResource.ts keeps only the
useState/useEffect wrapper (createCachedResource) and re-exports the rest
for existing callers. api/client.ts now reads dropCachesAfterWrite from
utils/resourceCache.ts — no more api -> hooks import.
